+++ CHANGELOG.rst
@@ -1,3 +1,16 @@
+Version 1.6.0 (30-04-2015)
+===========================================================
+
+*   Feature: Added a new range of scan modes that uses hash
+    instead of just filename or size to figure out the correct file
+*   Feature: Added support for dry-run with both text and
+    json output format
+*   Feature: Added support for scanning a single path
+*   Feature: Added commandline option to create base configuration
+    file.
+*   Bugfix: Trying to make sure torrent is added to rtorrent before
+    deleting it.
+
 Version 1.5.2 (19-03-2015)
 ===========================================================